requirements.txt
asgiref==3.5.2
Django==4.0.5
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-extensions==3.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
pydot==1.4.2
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
